# Conflicts:
#	companion/src/helpers.cpp
#	companion/src/helpers.h
#	companion/src/translations/companion_da.ts
#	companion/src/translations/companion_sv.ts
#	radio/src/datastructs.h
#	radio/src/fonts/lvgl/lv_font_noto_cn_24.c
#	radio/src/fonts/lvgl/lv_font_noto_tw_24.c
#	radio/src/gui/colorlcd/LvglWrapper.cpp
#	radio/src/gui/colorlcd/model_logical_switches.cpp
#	radio/src/gui/colorlcd/model_mixes.cpp
#	radio/src/gui/colorlcd/output_edit.cpp
#	radio/src/gui/colorlcd/screen_setup.cpp
#	radio/src/gui/colorlcd/special_functions.cpp
#	radio/src/gui/colorlcd/view_about.cpp
#	radio/src/gui/colorlcd/view_logical_switches.h
#	radio/src/lua/api_general.cpp
#	radio/src/storage/modelslist.cpp
#	radio/src/switches.cpp
#	radio/src/translations/cn.h
#	radio/src/translations/da.h
#	radio/src/translations/de.h
#	radio/src/translations/en.h
#	radio/src/translations/pl.h
#	radio/src/translations/se.h
#	radio/src/translations/tw.h

+++ companion/src/firmwares/radiodata.cpp
@@ -157,7 +157,7 @@ bool RadioData::deleteLabel(int index)
 
 bool RadioData::renameLabel(QString from, QString to)
 {
-  bool success = true;  
+  bool success = true;
   QString csvFrom = escapeCSV(from);
   QString csvTo = escapeCSV(to);
   int lengthdiff = csvTo.size() - csvFrom.size();
@@ -218,7 +218,7 @@ void RadioData::swapLabel(int indFrom, int indTo)
 
 bool RadioData::addLabelToModel(int index, QString label)
 {
-  if (index >= models.size()) return false;
+  if ((unsigned int)index >= models.size()) return false;
   label = escapeCSV(label);
 
   char *modelLabelCsv = models[index].labels;
@@ -237,7 +237,7 @@ bool RadioData::addLabelToModel(int index, QString label)
 
 bool RadioData::removeLabelFromModel(int index, QString label)
 {
-  if (index >= models.size()) return false;
+  if ((unsigned int)index >= models.size()) return false;
 
   QStringList lbls = fromCSV(QString::fromUtf8(models[index].labels));
   if (lbls.indexOf(label) >= 0) {
